chore!: remove activity filter code
