+++ services/wallet/activity/session_service_test.go
@@ -6,8 +6,9 @@ import (
 	"testing"
 
 	eth "github.com/ethereum/go-ethereum/common"
+
+	ac "github.com/status-im/status-go/services/wallet/activity/common"
 	"github.com/status-im/status-go/services/wallet/common"
-	"github.com/status-im/status-go/services/wallet/transfer"
 	"github.com/status-im/status-go/transactions"
 
 	"github.com/stretchr/testify/require"
@@ -21,7 +22,7 @@ func TestService_IncrementalUpdateOnTop(t *testing.T) {
 	allAddresses, pendings, ch, cleanup := setupTransactions(t, state, transactionCount, []transactions.TestTxSummary{{DontConfirm: true, Timestamp: transactionCount + 1}})
 	defer cleanup()
 
-	sessionID := state.service.StartFilterSession(allAddresses, allNetworksFilter(), Filter{}, 5, V2)
+	sessionID := state.service.StartFilterSession(allAddresses, allNetworksFilter(), Filter{Types: allActivityTypesFilter(), Statuses: allActivityStatusesFilter(), Assets: allTokensFilter()}, 5, V2)
 	require.Greater(t, sessionID, SessionID(0))
 	defer state.service.StopFilterSession(sessionID)
 
@@ -45,11 +46,11 @@ func TestService_IncrementalUpdateOnTop(t *testing.T) {
 
 		// Check the new transaction data
 		newTx := payload.Activities[0]
-		require.Equal(t, PendingTransactionPT, newTx.payloadType)
+		require.Equal(t, ac.PendingTransactionPT, newTx.payloadType)
 		// We don't keep type in the DB
 		require.Equal(t, (*int64)(nil), newTx.transferType)
-		require.Equal(t, SendAT, newTx.activityType)
-		require.Equal(t, PendingAS, newTx.activityStatus)
+		require.Equal(t, ac.SendAT, newTx.activityType)
+		require.Equal(t, ac.PendingAS, newTx.activityStatus)
 		require.Equal(t, exp.ChainID, newTx.transaction.ChainID)
 		require.Equal(t, exp.ChainID, *newTx.chainIDOut)
 		require.Equal(t, (*common.ChainID)(nil), newTx.chainIDIn)
@@ -62,17 +63,17 @@ func TestService_IncrementalUpdateOnTop(t *testing.T) {
 		require.Equal(t, exp.Timestamp, uint64(newTx.timestamp))
 		require.Equal(t, exp.Symbol, *newTx.symbolOut)
 		require.Equal(t, (*string)(nil), newTx.symbolIn)
-		require.Equal(t, &Token{
-			TokenType: Native,
+		require.Equal(t, &ac.Token{
+			TokenType: ac.Native,
 			ChainID:   5,
 		}, newTx.tokenOut)
-		require.Equal(t, (*Token)(nil), newTx.tokenIn)
+		require.Equal(t, (*ac.Token)(nil), newTx.tokenIn)
 		require.Equal(t, (*eth.Address)(nil), newTx.contractAddress)
 
 		// Check the order of the following transaction data
-		require.Equal(t, SimpleTransactionPT, payload.Activities[1].payloadType)
+		require.Equal(t, ac.SimpleTransactionPT, payload.Activities[1].payloadType)
 		require.Equal(t, int64(transactionCount), payload.Activities[1].timestamp)
-		require.Equal(t, SimpleTransactionPT, payload.Activities[2].payloadType)
+		require.Equal(t, ac.SimpleTransactionPT, payload.Activities[2].payloadType)
 		require.Equal(t, int64(transactionCount-1), payload.Activities[2].timestamp)
 	}, nil)
 
@@ -226,76 +227,6 @@ func TestService_IncrementalUpdateFetchWindowNoReset(t *testing.T) {
 	require.Equal(t, 1, eventActivityDoneCount)
 }
 
-// Simulate and validate a multi-step user flow that was also a regression in the original implementation
-func TestService_FilteredIncrementalUpdateResetAndClear(t *testing.T) {
-	state := setupTestService(t)
-	defer state.close()
-
-	transactionCount := 5
-	allAddresses, pendings, ch, cleanup := setupTransactions(t, state, transactionCount, []transactions.TestTxSummary{{DontConfirm: true, Timestamp: transactionCount + 1}})
-	defer cleanup()
-
-	// Generate new transaction for step 5
-	newOffset := transactionCount + 2
-	newTxs, newFromTrs, newToTrs := transfer.GenerateTestTransfers(t, state.service.db, newOffset, 1)
-	allAddresses = append(append(allAddresses, newFromTrs...), newToTrs...)
-
-	// 1. User visualizes transactions for the first time
-	sessionID := state.service.StartFilterSession(allAddresses, allNetworksFilter(), Filter{}, 4, V2)
-	require.Greater(t, sessionID, SessionID(0))
-	defer state.service.StopFilterSession(sessionID)
-
-	validateFilteringDone(t, ch, 4, nil, nil)
-
-	// 2. User applies a filter for pending transactions
-	err := state.service.UpdateFilterForSession(sessionID, Filter{Statuses: []Status{PendingAS}}, 4)
-	require.NoError(t, err)
-
-	filterResponseCount := validateFilteringDone(t, ch, 0, nil, nil)
-
-	// 3. A pending transaction is added
-	exp := pendings[0]
-	err = state.pendingTracker.StoreAndTrackPendingTx(&exp)
-	require.NoError(t, err)
-
-	vFn := getValidateSessionUpdateHasNewOnTopFn(t)
-	pendingTransactionUpdate, sessionUpdatesCount := validateSessionUpdateEvent(t, ch, &filterResponseCount, 1, vFn)
-
-	// 4. User resets the view and the new pending transaction has the new flag
-	err = state.service.ResetFilterSession(sessionID, 2)
-	require.NoError(t, err)
-
-	// Validate the reset data
-	eventActivityDoneCount := validateFilteringDone(t, ch, 1, func(payload FilterResponse) {
-		require.True(t, payload.Activities[0].isNew)
-		require.Equal(t, int64(transactionCount+1), payload.Activities[0].timestamp)
-	}, nil)
-
-	require.Equal(t, 1, pendingTransactionUpdate)
-	require.Equal(t, 1, filterResponseCount)
-	require.Equal(t, 1, sessionUpdatesCount)
-	require.Equal(t, 1, eventActivityDoneCount)
-
-	// 5. A new transaction is downloaded
-	transfer.InsertTestTransfer(t, state.service.db, newTxs[0].To, &newTxs[0])
-
-	// 6. User clears the filter and only the new transaction should have the new flag
-	err = state.service.UpdateFilterForSession(sessionID, Filter{}, 4)
-	require.NoError(t, err)
-
-	eventActivityDoneCount = validateFilteringDone(t, ch, 4, func(payload FilterResponse) {
-		require.True(t, payload.Activities[0].isNew)
-		require.Equal(t, int64(newOffset), payload.Activities[0].timestamp)
-		require.False(t, payload.Activities[1].isNew)
-		require.Equal(t, int64(newOffset-1), payload.Activities[1].timestamp)
-		require.False(t, payload.Activities[2].isNew)
-		require.Equal(t, int64(newOffset-2), payload.Activities[2].timestamp)
-		require.False(t, payload.Activities[3].isNew)
-		require.Equal(t, int64(newOffset-3), payload.Activities[3].timestamp)
-	}, nil)
-	require.Equal(t, 1, eventActivityDoneCount)
-}
-
 // Test the different session-related endpoints in a multi-threaded environment
 func TestService_MultiThread(t *testing.T) {
 	state := setupTestService(t)
@@ -321,9 +252,6 @@ func TestService_MultiThread(t *testing.T) {
 				go func() {
 					defer subwg.Done()
 					var suberr error
-
-					suberr = state.service.UpdateFilterForSession(sessionID, Filter{}, 5)
-					require.NoError(t, suberr)
 
 					suberr = state.service.ResetFilterSession(sessionID, 5)
 					require.NoError(t, suberr)
