--- services/wallet/activity/txDetails.sql
@@ -1,14 +1,0 @@
--- Query searches for additional details of a transaction
-
-SELECT
-	tx_hash,
-	blk_number,
-	network_id,
-	account_nonce,
-	tx,
-	contract_address,
-	base_gas_fee
-FROM
-	transfers
-WHERE
-	hash = ?
